Fixed loading images causing crashes!!!
Figuring out the issue was 99% of the battle,, turns out parts of the image loading function runs asynchronously, and if it wasn't done loading before it rendered it would crash. epic.
The only image related crash I have yet to fix is if theres over 30mb(?) of image data in VRAM it crashes.